Remove jcenter from the build
Currently jcenter returns 403 Forbidden. Also, the repository
is not up to date, an older version of C* (3.11.5) had to be used
in management-api-server, whereas all the other modules used
3.11.10.

This commit completely removes jcenter from the build. Ant
plugin is used to fetch C* tarball from Apache.

The approach taken here is far from ideal, it doesn't
use local .m2 cache, it caches the tarball in
management-api-server/.cassandra-bin dir.

This commit is a workaround to have the build running until
a better solutions is figured out.