Fix the Channel Pipeline initializer for CC4 Agent
This creates a ConnectionFactory wrapper that overrides
the Netty Channel class to not have a class cast
exception when retrieving the remoteAddress(). In
newer versions of Netty, this causes startup problems
with Management API. This same strategy is being
done for the C* 5.0 agent as well as the HCD CC5 agent.