Fix HCD Dispatcher.processRequest method signature
This adds the correct reflection method signature
lookup for Comverged Core 4. Previous versions were
still using the "processInit" method which is removed
in later CC4 versions. The "processRequest" method
signature fixed in this patch seems to be the correct
one and the one that should have been used all along.